fix: Redis config is not read from application.yaml
