+++ src/main/java/faang/school/postservice/config/redis/RedisConfiguration.java
@@ -1,0 +1,38 @@
+package faang.school.postservice.config.redis;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.serializer.GenericToStringSerializer;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+@Configuration
+@RequiredArgsConstructor
+public class RedisConfiguration {
+    private final RedisConfig redisConfig;
+
+    @Bean
+    @Qualifier("user-ban")
+    public RedisTemplate<String, Long> userBanRedisTemplate(RedisConnectionFactory connectionFactory) {
+        RedisTemplate<String, Long> template = new RedisTemplate<>();
+        template.setConnectionFactory(connectionFactory);
+
+        template.setKeySerializer(new StringRedisSerializer());
+        template.setValueSerializer(new GenericToStringSerializer<>(Long.class));
+        template.setHashKeySerializer(new StringRedisSerializer());
+        template.setHashValueSerializer(new GenericToStringSerializer<>(Long.class));
+        template.afterPropertiesSet();
+
+        return template;
+    }
+
+    @Bean
+    @Qualifier("user-ban")
+    ChannelTopic userBanTopic() {
+        return new ChannelTopic(redisConfig.getChannels().get("user_ban").name());
+    }
+}
